refactor: application내의 jwt의존성 adapter로 이동
