+++ src/main/java/yapp/buddycon/app/auth/adapter/JwtTokenCreator.java
@@ -1,9 +1,11 @@
-package yapp.buddycon.app.auth.application.service;
+package yapp.buddycon.app.auth.adapter;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import yapp.buddycon.app.auth.application.service.Token;
+import yapp.buddycon.app.auth.application.service.TokenCreator;
 import yapp.buddycon.app.user.domain.User;
 
 import java.util.Date;
