+++ src/main/java/yapp/buddycon/app/auth/adapter/JwtTokenProvider.java
@@ -1,9 +1,13 @@
-package yapp.buddycon.app.auth.application.service;
+package yapp.buddycon.app.auth.adapter;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import yapp.buddycon.app.auth.application.port.out.CacheStorage;
+import yapp.buddycon.app.auth.application.service.Time;
+import yapp.buddycon.app.auth.application.service.Token;
+import yapp.buddycon.app.auth.application.service.TokenCreator;
+import yapp.buddycon.app.auth.application.service.TokenProvider;
 import yapp.buddycon.app.user.domain.User;
 
 import java.util.Date;
